Split up Streamable trait
This commit splits up the Streamable trait into
a StreamReader and a StreamWriter trait.
I've also decided to remove the built-in support for
reading and writing String, Vec<T>, and HashMap<T, V>
since implementations of these are highly opinionated and
probably won't be of much use for consumers of this library.